角色授权优化，代码瘦身

+++ sodo-housekeeper/src/main/java/cool/sodo/housekeeper/service/impl/RoleToMenuServiceImpl.java
@@ -5,7 +5,9 @@ import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import cool.sodo.common.domain.RoleToMenu;
 import cool.sodo.common.entity.ResultEnum;
 import cool.sodo.common.exception.SoDoException;
-import cool.sodo.housekeeper.mapper.RoleToMenuMapper;
+import cool.sodo.common.mapper.CommonRoleToMenuMapper;
+import cool.sodo.common.service.impl.CommonRoleToMenuServiceImpl;
+import cool.sodo.common.util.StringUtil;
 import cool.sodo.housekeeper.service.RoleToMenuService;
 import org.springframework.stereotype.Service;
 
@@ -13,10 +15,10 @@ import javax.annotation.Resource;
 import java.util.List;
 
 @Service
-public class RoleToMenuServiceImpl implements RoleToMenuService {
+public class RoleToMenuServiceImpl extends CommonRoleToMenuServiceImpl implements RoleToMenuService {
 
     @Resource
-    private RoleToMenuMapper roleToMenuMapper;
+    private CommonRoleToMenuMapper roleToMenuMapper;
 
     @Override
     public void insert(List<RoleToMenu> roleToMenuList) {
@@ -29,12 +31,58 @@ public class RoleToMenuServiceImpl implements RoleToMenuService {
     }
 
     @Override
+    public void deleteByRole(String roleId) {
+
+        LambdaQueryWrapper<RoleToMenu> roleToMenuLambdaQueryWrapper = Wrappers.lambdaQuery();
+        roleToMenuLambdaQueryWrapper.eq(RoleToMenu::getRoleId, roleId);
+        if (roleToMenuMapper.delete(roleToMenuLambdaQueryWrapper) < 0) {
+            throw new SoDoException(ResultEnum.SERVER_ERROR, "删除 RoleToMenu 失败！");
+        }
+    }
+
+    @Override
     public void deleteByRole(List<String> roleIdList) {
 
+        if (StringUtil.isEmpty(roleIdList)) {
+            return;
+        }
         LambdaQueryWrapper<RoleToMenu> roleToMenuLambdaQueryWrapper = Wrappers.lambdaQuery();
         roleToMenuLambdaQueryWrapper.in(RoleToMenu::getRoleId, roleIdList);
         if (roleToMenuMapper.delete(roleToMenuLambdaQueryWrapper) < 0) {
             throw new SoDoException(ResultEnum.SERVER_ERROR, "删除 RoleToMenu 失败！");
         }
+    }
+
+    @Override
+    public void deleteByMenu(String menuId) {
+
+        LambdaQueryWrapper<RoleToMenu> roleToMenuLambdaQueryWrapper = Wrappers.lambdaQuery();
+        roleToMenuLambdaQueryWrapper.eq(RoleToMenu::getMenuId, menuId);
+        if (roleToMenuMapper.delete(roleToMenuLambdaQueryWrapper) < 0) {
+            throw new SoDoException(ResultEnum.SERVER_ERROR, "删除 RoleToMenu 失败！");
+        }
+    }
+
+    @Override
+    public void deleteByMenu(List<String> menuIdList) {
+
+        if (StringUtil.isEmpty(menuIdList)) {
+            return;
+        }
+        LambdaQueryWrapper<RoleToMenu> roleToMenuLambdaQueryWrapper = Wrappers.lambdaQuery();
+        roleToMenuLambdaQueryWrapper.in(RoleToMenu::getMenuId, menuIdList);
+        if (roleToMenuMapper.delete(roleToMenuLambdaQueryWrapper) < 0) {
+            throw new SoDoException(ResultEnum.SERVER_ERROR, "删除 RoleToMenu 失败！");
+        }
+    }
+
+    @Override
+    public List<String> listRoleToMenuMenuIdByRole(List<String> roleIdList) {
+        return super.listRoleToMenuMenuIdByRole(roleIdList);
+    }
+
+    @Override
+    public List<String> listRoleToMenuMenuIdByRole(String roleId) {
+        return super.listRoleToMenuMenuIdByRole(roleId);
     }
 }
